modify the structur
